more giblets

+++ src/game/entities/objects/projectiles/bullet.rs
@@ -12,7 +12,7 @@ fn bullet_new(parameters: &ObjectSummonParameters) -> GameObject {
     .add_controller(CONTROLLER_TYPE_UPDATE,Box::new(ProjectileController::new(parameters.direction, parameters.speed)))
     .add_controller(CONTROLLER_TYPE_WORLD_COLLIDE, Box::new(DestroyOnEvent::new()))
     .add_controller(CONTROLLER_TYPE_OBJECT_COLLIDE, Box::new(DestroyOnHostileCollision::new()))
-    .add_controller(CONTROLLER_TYPE_DESTROYED, Box::new(SpawnGibletsOnDeathController::new(GIBLET_SPLINTER as u8, 1, 3, vec!(Color::new(0.843, 0.451, 0.333, 1.0), Color::new(0.863, 0.961, 1.0, 1.0)))))
+    .add_controller(CONTROLLER_TYPE_DESTROYED, Box::new(SpawnGibletsOnDeathController::new(GIBLET_SPLINTER, 1, 3, vec!(Color::new(0.843, 0.451, 0.333, 1.0), Color::new(0.863, 0.961, 1.0, 1.0)))))
     .build();
 }
